internal/tools: split runner helpers into semantic files
- Extract IO helper (safeReadAll) to runner_io.go
- Extract audit logging (writeAudit, appendAuditLog, moduleRoot) to runner_audit.go
- Extract redaction utilities to runner_redact.go
- Introduce ToolSpec in internal/tools/spec.go for local tests

Refs: https://github.com/hyperifyio/goagent/pull/10#discussion_r2288632449
